Move A/B test list to its own module
